Fix: 파일명(배너,헤더),type,경로 수정
