+++ src/types/mock-data-type.ts
@@ -1,5 +1,4 @@
 export interface Track {
-
   album: Album
   artists: Artist[]
   disc_number: number
@@ -19,22 +18,6 @@ export interface Track {
   is_local: boolean
 }
 
-export interface Album {
-  album_type: string
-  total_tracks: number
-  external_urls: ExternalUrls
-  href: string
-  id: string
-  images: Image[]
-  name: string
-  release_date: Date
-  release_date_precision: string
-  type: string
-  uri: string
-  artists: Artist[]
-  is_playable: boolean
-}
-
 export interface Artist {
   external_urls: ExternalUrls
   href: string
@@ -48,39 +31,38 @@ export interface Artist {
 export interface ExternalUrls {
   spotify: string
 
-  album: Album;
-  artists: Artist[];
-  disc_number: number;
-  duration_ms: number;
-  explicit: boolean;
-  id: string;
-  is_playable?: boolean;
-  name: string;
-  popularity: number;
-  preview_url: string;
-  track_number: number;
-  type: string;
-  is_local: boolean;
+  album: Album
+  artists: Artist[]
+  disc_number: number
+  duration_ms: number
+  explicit: boolean
+  id: string
+  is_playable?: boolean
+  name: string
+  popularity: number
+  preview_url: string
+  track_number: number
+  type: string
+  is_local: boolean
 }
 
 export interface Album {
-  album_type: string;
-  total_tracks: number;
-  id: string;
-  images: Image[];
-  name: string;
-  release_date: string;
-  release_date_precision: string;
-  type: string;
-  artists: Artist[];
-  is_playable?: boolean;
+  album_type: string
+  total_tracks: number
+  id: string
+  images: Image[]
+  name: string
+  release_date: string
+  release_date_precision: string
+  type: string
+  artists: Artist[]
+  is_playable?: boolean
 }
 
 export interface Artist {
-  id: string;
-  name: string;
-  type: string;
-
+  id: string
+  name: string
+  type: string
 }
 
 export interface Image {
@@ -89,11 +71,9 @@ export interface Image {
   width: number
 }
 
-
 export interface ExternalIDS {
   isrc: string
 }
-
 
 ///
 export interface FollowArtist {
@@ -101,15 +81,13 @@ export interface FollowArtist {
 }
 
 export interface Artists {
-
   href: string
- 
-  limit: number;
-  next: null;
-  cursors: Cursors;
-  total: number;
-  items: Item[];
 
+  limit: number
+  next: null
+  cursors: Cursors
+  total: number
+  items: Item[]
 }
 
 export interface Cursors {
@@ -117,7 +95,6 @@ export interface Cursors {
 }
 
 export interface Item {
-
   external_urls: ExternalUrls
   followers: Followers
   genres: string[]
@@ -166,14 +143,13 @@ export interface Tracks {
   href: string
   total: number
 
-  followers: Followers;
-  genres: string[];
-  id: string;
-  images: Image[];
-  name: string;
-  popularity: number;
-  type: Type;
-
+  followers: Followers
+  genres: string[]
+  id: string
+  images: Image[]
+  name: string
+  popularity: number
+  type: Type
 }
 // export interface Artist extends Partial<Item> {
 //   external_urls: ExternalUrls;
